Code review (removing if-else from main)

+++ price-calc/src/main/java/com/epam/brest/selector/PriceSelector.java
@@ -1,14 +1,13 @@
 package com.epam.brest.selector;
 
 import java.math.BigDecimal;
-import java.math.BigInteger;
 import java.util.Map;
 import java.util.SortedSet;
 import java.util.TreeSet;
 
 public class PriceSelector {
 
-    public BigDecimal selectPriceValue(Map<Integer, BigDecimal> valueMap, BigInteger targetValue) {
+    public BigDecimal selectPriceValue(Map<Integer, BigDecimal> valueMap, BigDecimal targetValue) {
         SortedSet<Integer> sortedKeys = new TreeSet<>(valueMap.keySet());
         Integer resultKey = sortedKeys.first();
         for (Integer key : sortedKeys) {
